+++ echo_server/utils.h
@@ -3,6 +3,6 @@
 
 int send_all(int socket_fd, char *buffer, int length);
 int recv_all(int socket_fd, char *buffer, int length);
-int valid_port(const char *port_str, long *port);
+int valid_port(const char *port_str, int *port);
 
 #endif
